pyproject.toml
[project]
name = "imaginet-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]",
    "fastmcp",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.4.0",
]
